play chess
add chess is a 64 by 64 matrix
a matrix is a string that holds 2 dimensions of values
add white
add black

on which chess is played
chess is a game
a game is something that you play
a chess board is a 8 by 8 bit matrix in which 1 is black and 0 is white

a chess board is referred to as a board
the chess board has white and black values depending on the value of the
array and the element of the array
arrays are called columns and the entries of the arrays with similar
second values are rows
the elements of the columns are referred to as
1 equals A
2 equals B
3 equals C
4 equals D
5 equals E
6 equals F

8 equals H
1,1 is referred to as A1
1,2 is referred to as A2
1,3 is referred to as A3
1,4 is referred to as A4
1,5 is referred to as A5
and so on.
2,1 is referred to as B1
2,2 is referred to as B2
and so on
and so on.
add rook
add bishop
add knight
add queen
add king

the white king starts at e1
e1 is E1
e2 is E2
and so on
white queen starts at d1

there are 4 bishops
there are 4 knights
there are 4 rooks
there are 2 queens
there are 2 kings

one queen is black, 1 queen is white
2 bishops are black, 2 bishops are white
2 knights are black, 2, knights are white

the objective is to kill the king,
without getting killed.

cince chess is a game, you will play showing a 8 by 8 matrix in which
the elements of the matrix are black and white squares

play chess means run chess
run chess wait 100